# src/api.yaml
---
openapi: 3.0.0
info:
    title: ToDo API
    description: ToDo
    contact:
        email: johannes@sch.8.de
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 1.0.0
servers:
    - url: https://virtserver.swaggerhub.com/johannes3/ToDo/1.0.0
      description: SwaggerHub API Auto Mocking
    - url: https://sch-8.de/api/todo
      description: Test server (node.js)
tags:
    - name: tasks
      description: Tasks to be done
    - name: users
      description: Users that are assigned to tasks
paths:
    /tasks:
        get:
            tags:
                - tasks
            summary: get list of tasks
            description: |
                Get all tasks currently defined
            operationId: getTasks
            responses:
                '200':
                    description: list of all tasks
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/taskItem'
                '400':
                    description: bad input parameter
        post:
            tags:
                - tasks
            summary: add a task to the task list
            description: Adds an item to the system
            operationId: addTask
            requestBody:
                description: task to add
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/taskItem'
            responses:
                '201':
                    description: task created
                '400':
                    description: invalid input, object invalid
                '500':
                    description: Internal error, too many tasks

components:
    schemas:
        taskItem:
            required:
                - description
                - dueDate
                - name
            #- user
            type: object
            properties:
                id:
                    type: number
                    example: 32
                name:
                    type: string
                    example: 'Create an OpenApi defintion'
                description:
                    type: string
                    example: 'Use swagger and or Postman'
                dueDate:
                    type: string
                    format: date-time
                    example: '2016-08-29T09:12:33.001Z'
                user:
                    type: number
                    example: 55
        userItem:
            required:
                - name
            type: object
            properties:
                name:
                    type: string
                    example: Johannes
                homePage:
                    type: string
                    format: url
                    example: https://www.acme-corp.com
                phone:
                    type: string
                    example: 408-867-5309